refactor the presto check operator into a generalized check operator
